+++ cmd/pytestsim/main.go
@@ -1,0 +1,14 @@
+// This code is released under the MIT License
+// Copyright (c) 2023 Marco Molteni and the timeit contributors.
+
+package main
+
+import (
+	"os"
+
+	"github.com/marco-m/timeit/pkg/pytestsim"
+)
+
+func main() {
+	os.Exit(pytestsim.Main())
+}